Extract leadership profiles to a new block
